requirements.txt
fastapi
uvicorn
pydantic
google-analytics-data
google-auth
google-api-python-client
openai
python-dotenv
pandas
